Adding newer version of code to betaBranch on GitHub repository
