+++ src/ua/lviv/iot/home/electrical/devices/models/Fridge.java
@@ -1,4 +1,4 @@
-package ua.lviv.iot.electrical.devices.models;
+package ua.lviv.iot.home.electrical.devices.models;
 
 public class Fridge extends HomeElectricalDevice {
